Moved Submissions API Crypto bits into IdentityModel
