+++ src/IdentityModel/IdentityTool/Startup.cs
@@ -1,6 +1,4 @@
-﻿using Biobanks.IdentityModel.Services;
-
-using Biobanks.IdentityTool.Commands.Runners;
+﻿using Biobanks.IdentityTool.Commands.Runners;
 
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
@@ -14,8 +12,6 @@ namespace Biobanks.IdentityTool
 
 
             services
-                .AddTransient<CryptoService>()
-
                 // Command Runners
                 .AddTransient<GenerateId>();
         }
