+++ src/Submissions/Api/Api.csproj
@@ -36,6 +36,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\Data\Data.csproj" />
+    <ProjectReference Include="..\..\IdentityModel\IdentityModel\IdentityModel.csproj" />
     <ProjectReference Include="..\Core.AzureStorage\Core.AzureStorage.csproj" />
     <ProjectReference Include="..\Core\Core.csproj" />
   </ItemGroup>
