+++ src/Submissions/Api/Auth/Basic/BasicAuthHandler.cs
@@ -1,4 +1,12 @@
-﻿using System;
+﻿using Biobanks.Data;
+using Biobanks.IdentityModel.Extensions;
+
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http.Headers;
@@ -6,13 +14,6 @@ using System.Security.Claims;
 using System.Text;
 using System.Text.Encodings.Web;
 using System.Threading.Tasks;
-
-using Biobanks.Data;
-
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
 
 namespace Biobanks.Submissions.Api.Auth.Basic
 {
